// backend/src/types/modules/prijaveRepository.ts
export type PrijavaRow = {
  BROJ_PRIJAVE: number;
  DATUM_PRIJAVE: Date | null;
  SKOLSKA_GODINA: string;
  ID_KONKURSA: number | null;
  ID_KORISNIKA: number | null;
  ID_PROGRAMA: number | null;
  STATUS_PRIJAVE: string | null;
  KONKURSNI_ROK: string | null;
  JMBG: string | null;
  IME_PREZIME: string | null;
  SISTEMSKI_UPDATE: string | null;
};